# server/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RunMe API",
    "description": "REST API for the RunMe mobile app (auth, loans, grants, wallet, admin). Use **Authorize** with a Bearer token from `POST /auth/login` or `POST /auth/register`.",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "/", "description": "Current host" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the `accessToken` value from login/register (without the word Bearer)."
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 1 }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 }
        }
      },
      "TokenPair": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string" }
        }
      },
      "LogoutRequest": {
        "type": "object",
        "properties": {
          "refreshToken": { "type": "string", "description": "Optional: revokes refresh token server-side" }
        }
      },
...
      "AdminSummary": {
        "type": "object",
        "properties": {
          "usersCount": { "type": "integer" },
          "generatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Loan": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "borrowerId": { "type": "string", "format": "uuid" },
          "lenderId": { "type": "string", "format": "uuid", "nullable": true },
          "amount": { "type": "number" },
          "purpose": { "type": "string", "enum": ["rent", "food", "emergency"] },
          "durationDays": { "type": "integer" },
          "interestRate": { "type": "number" },
          "status": { "type": "string", "enum": ["pending", "funded", "repaid"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "audience": { "type": "string", "enum": ["public", "friendsOnly"] },
          "reason": { "type": "string", "nullable": true },
          "proofFileUrl": { "type": "string", "nullable": true },
          "repaidAmount": { "type": "number" }
        }
      },
      "LoanCreate": {
        "type": "object",
        "required": ["amount", "purpose", "durationDays", "audience"],
        "properties": {
          "amount": { "type": "number", "minimum": 0 },
          "purpose": { "type": "string", "enum": ["rent", "food", "emergency"] },
          "durationDays": { "type": "integer", "minimum": 7, "maximum": 365 },
          "audience": { "type": "string", "enum": ["public", "friendsOnly"] },
          "reason": { "type": "string" },
          "proofFileUrl": { "type": "string", "format": "uri", "nullable": true }
        }
      },
...
      "Grant": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "story": { "type": "string" },
          "goalNaira": { "type": "number" },
          "raisedNaira": { "type": "number" },
          "category": { "type": "string", "enum": ["studentStory", "urgentNeed", "education"] },
          "studentName": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "attachmentUrls": { "type": "array", "items": { "type": "string" } },
          "isUrgent": { "type": "boolean" }
        }
      },
      "GrantCreate": {
        "type": "object",
        "required": ["title", "story", "goalNaira", "category", "studentName"],
        "properties": {
          "title": { "type": "string" },
          "story": { "type": "string" },
          "goalNaira": { "type": "number", "minimum": 0 },
          "category": { "type": "string", "enum": ["studentStory", "urgentNeed", "education"] },
          "studentName": { "type": "string" },
          "isUrgent": { "type": "boolean" },
          "attachmentUrls": { "type": "array", "items": { "type": "string", "format": "uri" } }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenPair" } } }
          },
          "400": { "description": "Validation error" },
          "409": { "description": "Email already in use" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenPair" } } }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenPair" } } }
          },
          "401": { "description": "Invalid refresh token" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout (blacklist access token; optional refresh revoke)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LogoutRequest" } }
          }
        },
        "responses": {
          "204": { "description": "No content" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/admin/summary": {
      "get": {
        "tags": ["Admin"],
        "summary": "Admin dashboard summary",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminSummary" } } }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden (not admin)" }
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "List users",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 100 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/admin/loans": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all loans",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/admin/grants": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all grants",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/admin/audit": {
      "get": {
        "tags": ["Admin"],
        "summary": "Audit log",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/loans": {
      "get": {
        "tags": ["Loans"],
        "summary": "List loans",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "borrowerId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "funded", "repaid"] } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/Loan" } },
                    "nextCursor": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Loans"],
        "summary": "Create loan request",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoanCreate" } }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Loan" } } }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/grants": {
      "get": {
        "tags": ["Grants"],
        "summary": "List grants",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string", "enum": ["studentStory", "urgentNeed", "education"] } },
          { "name": "urgentOnly", "in": "query", "schema": { "type": "boolean" } },
          { "name": "excludeBorrowerId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/Grant" } },
                    "nextCursor": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Grants"],
        "summary": "Create grant request",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/GrantCreate" } }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Grant" } } }
          }
        }
      }
    },
...
  }
}